+++ CHANGELOG.md
@@ -15,10 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed a bug in `subject` (extra separator) and in `receivers` (long strings) parameters in Outlook connector. 
+- Fixed issue with credentials handling in `VidClub` source class
+- Fixed issue with missing arguments in `VidClubToDF` task class
+
 
 ### Changed
 - Genesys API call method and the name changed from `genesys_generate_exports` to `genesys_api_connection`. 
 - Added `GET` connection inside the method `genesys_api_connection`.
+- Changed looping structure for API calls in `VidClub` source class to use time intervals
+- Changed `VidClubToDF` task class to use total_load function from source
 
 ### Removed
 - Removed methods never used in production: `get_analitics_url_report`, `get_all_schedules_job`, `schedule_report`,